% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiper_glm.R
\name{hiper_glm}
\alias{hiper_glm}
\title{Sequence optimize}
\usage{
hiper_glm(design, outcome, model = "linear")
}
\arguments{
\item{design}{matrix, the design matrix X for the predictors.}

\item{outcome}{vector, the output Y for the response.}

\item{model}{the model to be used. The default is set as 'linear' to use the linear model;
otherwise use 'logit' for}
}
\value{
A list including the information of the fitted model
}
\description{
Do sequence optimize
}
